Implement classic State Machine.
